requirements.txt
asttokens==3.0.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brlapi==0.8.5
Brotli==1.1.0
ccsm==0.9.14.2
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
click==8.1.6
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
compizconfig-python==0.9.14.2
configobj==5.0.8
cryptography==41.0.7
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.3.2
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
Deprecated==1.2.14
deprecation==2.0.7
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
executing==2.2.0
eyed3==0.9.7
filelock==3.13.1
filetype==1.2.0
Flask==3.1.1
Flask-PyMongo==3.0.1
gbinder-python==1.1.2
html2text==2024.2.26
httplib2==0.20.4
idna==3.6
ifaddr==0.2.0
IMDbPY==2021.4.18
img2pdf==0.5.1
importlib-metadata==4.12.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.2.1
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==24.3.1
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libevdev==0.5
louis==3.29.0
lxml==5.2.1
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.2.0
multiprocess==0.70.16
mutagen==1.46.0
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
onboard==1.4.1
packaging==24.0
PAM==0.4.2
parso==0.8.4
pdfarranger==1.10.1
pexpect==4.9.0
pikepdf==8.7.1
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasyncore==1.0.2
pycairo==1.25.1
pycparser==2.22
pycryptodomex==3.20.0
pycups==2.0.1
pycurl==7.45.3
pyelftools==0.30
Pygments==2.17.2
PyGObject==3.48.2
PyICU==2.12
pyinotify==0.9.6
PyJWT==2.7.0
pymongo==4.13.2
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyparted==3.12.0
pypdf==5.7.0
pypng==0.20231004.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
python-apt==2.7.7+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-dotenv==1.1.1
python-gnupg==0.5.2
python-magic==0.4.27
python-xlib==0.33
pytz==2024.1
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.1
pyzmq==27.0.0
qrcode==7.4.2
repolib==2.2.1
requests==2.31.0
requests-file==1.5.1
rich==13.7.1
scour==0.38.2
SecretStorage==3.3.3
setproctitle==1.3.3
setuptools==68.1.2
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
systemd-python==235
tinycss2==1.2.1
tldextract==3.1.2
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ufw==0.36.2
Unidecode==1.3.8
urllib3==2.0.7
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
Werkzeug==3.1.3
wheel==0.42.0
wrapt==1.15.0
xdg==5
xkit==0.0.0
xlrd==2.0.1
yt-dlp==2024.4.9
zipp==1.0.0
